PSLA: rename PSLA paper
There is something funny about the way we name files. We tend to
name them so that when we run `ls` in our website, we can tell
all the papers apart. We should be naming them so that when
someone downloads our paper, they can easily find it in their
`Downloads` folder. Thus adding relevant information like author
or venue to the filename can be useful!